services/notification-service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
aiokafka==0.10.0
apscheduler==3.10.4
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.1
